+++ CHANGELOG.md
@@ -2,6 +2,7 @@ musicscan CHANGELOG
 ======================
 
 ## CURRENT
+ - [musicscan-36](https://github.com/cjcodeproj/musicscan/issues/36) Update some editing notes
  - [musicscan-52](https://github.com/cjcodeproj/musicscan/issues/52) Add new flags to RST documentation
  - [musicscan-40](https://github.com/cjcodeproj/musicscan/issues/40) Documentation: flags option flag
  - [musicscan-53](https://github.com/cjcodeproj/musicscan/issues/53) Album title is not sanitized